pyproject.toml
[project]
name = "contextosolver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastembed>=0.7.0",
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "nltk>=3.9.1",
    "omegaconf>=2.3.0",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.4",
    "pyfiglet>=1.0.2",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
    "scikit-learn>=1.6.1",
    "spacy>=3.8.6",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]

[tool.setuptools]
package-dir = {"" = "src"}
